+++ screens/EventScreen/EventScreen.tsx
@@ -190,7 +190,7 @@ export default class EventScreen extends React.Component<Props, State>{
                     : null
                   }
                   {this.state.canDelete ?
-                    <Button onPress={() => { this.delete() }} bordered style={styles.sliderButton}><Text>Delete Event</Text></Button>
+                    <Button onPress={() => { if (window.confirm('Are you sure you wish to delete this event?')) this.delete() }} bordered style={styles.sliderButton}><Text>Delete Event</Text></Button>
                     : null
                   }
                   <Text>{this.state.validationError}</Text>
