+++ screens/ResourceScreen/ResourceScreen.tsx
@@ -187,7 +187,7 @@ export default class GroupScreen extends React.Component<Props, State>{
                     : null
                   }
                   {this.state.canDelete ?
-                    <Button onPress={() => { this.delete() }} bordered style={styles.sliderButton}><Text>Delete Resource</Text></Button>
+                    <Button onPress={() => { if (window.confirm('Are you sure you wish to delete this resource?')) this.delete() }} bordered style={styles.sliderButton}><Text>Delete Resource</Text></Button>
                     : null
                   }
                   <Text>{this.state.validationError}</Text>
